refactor: PostProcessStage
