+++ src/PostProcessStage/BrightnessStage.ts
@@ -1,0 +1,13 @@
+import { PostProcessStageLibrary } from "cesium";
+
+import { createPostProcessStage } from "../core/PostProcessStage";
+
+export const BrightnessStage = createPostProcessStage<{
+  brightness?: number;
+}>({
+  name: "BrightnessStage",
+  props: ["brightness"],
+  create: () => PostProcessStageLibrary.createBrightnessStage(),
+});
+
+export default BrightnessStage;
