+++ src/PostProcessStage/Fxaa.ts
@@ -1,0 +1,9 @@
+import { createPostProcessStage } from "../core/PostProcessStage";
+
+export const Fxaa = createPostProcessStage<{}>({
+  name: "Fxaa",
+  create: (props, collection) => collection.fxaa,
+  props: [],
+});
+
+export default Fxaa;
