+++ src/PostProcessStage/LensFlareStage.ts
@@ -1,0 +1,26 @@
+import { PostProcessStageLibrary } from "cesium";
+
+import { createPostProcessStage } from "../core/PostProcessStage";
+
+export const LensFlareStage = createPostProcessStage<{
+  dirtTexture?: any;
+  starTexture?: any;
+  intensity?: number;
+  distortion?: number;
+  ghostDispersal?: number;
+  haloWidth?: number;
+  earthRadius?: number;
+}>({
+  name: "LensFlareStage",
+  props: [
+    "dirtTexture",
+    "starTexture",
+    "intensity",
+    "distortion",
+    "ghostDispersal",
+    "haloWidth",
+    "earthRadius",
+  ],
+  create: () => PostProcessStageLibrary.createLensFlareStage(),
+});
+export default LensFlareStage;
