+++ src/PostProcessStage/NightVisionStage.ts
@@ -1,0 +1,11 @@
+import { PostProcessStageLibrary } from "cesium";
+
+import { createPostProcessStage } from "../core/PostProcessStage";
+
+export const NightVisionStage = createPostProcessStage<{}>({
+  name: "NightVisionStage",
+  props: [],
+  create: () => PostProcessStageLibrary.createNightVisionStage(),
+});
+
+export default NightVisionStage;
