+++ src/PostProcessStage/index.ts
@@ -1,9 +1,6 @@
-export {
-  PostProcessStage,
-  BlackAndWhiteStage,
-  BrightnessStage,
-  LensFlareStage,
-  Fxaa,
-  NightVisionStage,
-} from "./PostProcessStage";
-export { default } from "./PostProcessStage";
+export { PostProcessStage, default } from "./PostProcessStage";
+export { BlackAndWhiteStage } from "./BlackAndWhiteStage";
+export { BrightnessStage } from "./BrightnessStage";
+export { LensFlareStage } from "./LensFlareStage";
+export { Fxaa } from "./Fxaa";
+export { NightVisionStage } from "./NightVisionStage";
