+++ src/PostProcessStageComposite/Bloom.ts
@@ -1,0 +1,17 @@
+import { createPostProcessStage } from "../core/PostProcessStage";
+
+export const Bloom = createPostProcessStage<{
+  contrast?: number;
+  brightness?: number;
+  glowOnly?: boolean;
+  delta?: number;
+  sigma?: number;
+  stepSize?: number;
+}>({
+  name: "Bloom",
+  create: (props, collection) => (collection as any).bloom,
+  props: ["brightness", "contrast", "delta", "glowOnly", "sigma", "stepSize"],
+  noMount: true,
+});
+
+export default Bloom;
